// packages/backend.ai-ui/src/__generated__/BAIProjectVfolderSelectAstryxPaginatedQuery.graphql.ts
export type VFolderFilter = {
  AND?: ReadonlyArray<VFolderFilter> | null | undefined;
  NOT?: ReadonlyArray<VFolderFilter> | null | undefined;
  OR?: ReadonlyArray<VFolderFilter> | null | undefined;
  cloneable?: boolean | null | undefined;
  createdAt?: DateTimeFilter | null | undefined;
  host?: StringFilter | null | undefined;
  name?: StringFilter | null | undefined;
  status?: VFolderOperationStatusFilter | null | undefined;
  usageMode?: VFolderUsageModeFilter | null | undefined;
};
// ...
export type BAIProjectVfolderSelectAstryxPaginatedQuery$variables = {
  filter?: VFolderFilter | null | undefined;
  limit: number;
  offset: number;
  projectId: string;
};
export type BAIProjectVfolderSelectAstryxPaginatedQuery$data = {
  readonly projectVfolders: {
    readonly count: number;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly metadata: {
          readonly name: string;
        };
      };
    }>;
  } | null | undefined;
};
export type BAIProjectVfolderSelectAstryxPaginatedQuery = {
  response: BAIProjectVfolderSelectAstryxPaginatedQuery$data;
  variables: BAIProjectVfolderSelectAstryxPaginatedQuery$variables;
};